Изменения в UI. Добавлено сохранение и изменение время обновления данных

+++ Joystick/Joystick/Data/DataResponse.cs
@@ -2,8 +2,8 @@
 {
     public struct DataResponse
     {
-        public int XPosition { get; set; }
-        public int YPosition { get; set; }
+        public short XPosition { get; set; }
+        public short YPosition { get; set; }
         public bool IsEnableLights { get; set; }
     }
 }
